Checkpoint before assistant change: Restructure issue processing services for better organization and clarity
Move IssueSaveService and IssueTimeConverter to the issue subdirectory.
